requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
litellm==1.31.6
boto3==1.36.0
python-dotenv==1.0.1
streamlit==1.35.0
requests
streamlit